+++ tests/cypress/integration/upload.cy.js
@@ -11,32 +11,37 @@ describe("Upload a file", function () {
     //create a unique name
     const folderName = Math.random().toString(36).substring(2, 7);
     cy.addFolder(folderName);
-    cy.wait(11000);
+    cy.wait(5000);
     cy.searchFolder(folderName);
     cy.get(".media-content").contains(folderName).click({ force: true });
-    cy.wait(5000);
-    cy.get('[data-testid="upload-file"]').click();
+    cy.wait(3000);
 
-    cy.wait(11000);
+    //press upload button from folder
+    cy.get('[data-testid="upload-file"]').click({ force: true });
+    cy.wait(5000);
 
     //upload the fixture file
     cy.get(".upload-btn-wrapper")
       .find("input")
       .invoke("show")
       .selectFile(`cypress/fixtures/text-files/text-file.txt`);
-    cy.wait(20000);
-    cy.get(".upload-card > c-card-actions.hydrated > :nth-child(2)").click();
-    cy.wait(20000);
+    cy.wait(8000);
+    cy.get(".upload-card > c-card-actions.hydrated > :nth-child(2)").click({
+      force: true,
+    });
+    cy.wait(8000);
 
     //close the modal
-    cy.get(".link-underline").click();
+    cy.get(".link-underline").click({ force: true });
 
     cy.reload();
-    cy.wait(15000);
+    cy.wait(8000);
+
     //check if the file name is on the page
     cy.contains("text-file").should("exist");
 
-    //delete the file
+    //delete the file by checkbox
+
     cy.contains("text-file")
       .parent()
       .parent()
@@ -44,13 +49,12 @@ describe("Upload a file", function () {
       .eq(5)
       .find("button")
       .eq(2)
-      .click();
+      .click({ force: true });
+    cy.get("c-alert.hydrated > c-card-actions.hydrated > :nth-child(2)").click({
+      force: true,
+    });
 
-    //confirm deletion
-    cy.get(
-      "c-alert.hydrated > c-card-actions.hydrated > :nth-child(2)"
-    ).click();
-    cy.wait(15000);
+    cy.wait(8000);
     cy.contains("This folder has no content.").should("exist");
   });
 
@@ -59,9 +63,9 @@ describe("Upload a file", function () {
     //create a unique name
     const randomName = Math.random().toString(36).substring(2, 7);
     cy.addFolder(randomName);
-    cy.wait(10000);
+    cy.wait(5000);
     cy.get('[data-testid="upload-file"]').click({ force: true });
-    cy.wait(10000);
+    cy.wait(5000);
 
     //trying to avoid '0x0' effective width and height error and type into input field
     cy.contains("Folder name")
@@ -69,41 +73,35 @@ describe("Upload a file", function () {
       .find("input")
       .type(randomName, { force: true });
 
-    cy.get('[role="listbox"]').find("li").contains(randomName).click();
-    cy.wait(15000);
+    cy.get('[role="listbox"]')
+      .find("li")
+      .contains(randomName)
+      .click({ force: true });
+    cy.wait(8000);
 
     //   upload the fixture file
     cy.get(".upload-btn-wrapper")
       .find("input")
       .invoke("show")
       .selectFile("cypress/fixtures/text-files/text-file.txt");
-    cy.wait(20000);
-    cy.get(".upload-card > c-card-actions.hydrated > :nth-child(2)").click();
-    cy.wait(20000);
+    cy.wait(8000);
+    cy.get(".upload-card > c-card-actions.hydrated > :nth-child(2)").click({
+      force: true,
+    });
+    cy.wait(8000);
 
     //close the modal
-    cy.get(".link-underline").click();
+    cy.get(".link-underline").click({ force: true });
 
     //   check if the file name is on the page
     cy.reload();
-    cy.wait(15000);
+    cy.wait(8000);
     cy.contains("text-file").should("exist");
 
     //delete the file
-    cy.contains("text-file")
-      .parent()
-      .parent()
-      .find("td")
-      .eq(5)
-      .find("button")
-      .eq(2)
-      .click();
+    cy.deleteFileCheckbox("text-file");
 
-    //confirm deletion
-    cy.get(
-      "c-alert.hydrated > c-card-actions.hydrated > :nth-child(2)"
-    ).click();
-    cy.wait(10000);
+    cy.wait(5000);
     cy.contains("This folder has no content.").should("exist");
   });
 
@@ -111,9 +109,9 @@ describe("Upload a file", function () {
     //create a unique name
     const randomName = Math.random().toString(36).substring(2, 7);
     cy.addFolder(randomName);
-    cy.wait(10000);
+    cy.wait(5000);
     cy.get('[data-testid="upload-file"]').click({ force: true });
-    cy.wait(10000);
+    cy.wait(5000);
 
     //trying to avoid '0x0' effective width and height error and type into input field
     cy.contains("Folder name")
@@ -121,22 +119,25 @@ describe("Upload a file", function () {
       .find("input")
       .type(randomName, { force: true });
 
-    cy.get('[role="listbox"]').find("li").contains(randomName).click();
-    cy.wait(15000);
+    cy.get('[role="listbox"]')
+      .find("li")
+      .contains(randomName)
+      .click({ force: true });
+    cy.wait(8000);
 
     //   upload the fixture file
     cy.get(".upload-btn-wrapper")
       .find("input")
       .invoke("show")
       .selectFile("cypress/fixtures/text-files/text-file.txt");
-    cy.wait(10000);
+    cy.wait(5000);
 
     //try to upload the same file second time
     cy.get(".upload-btn-wrapper")
       .find("input")
       .invoke("show")
       .selectFile("cypress/fixtures/text-files/text-file.txt");
-    cy.wait(10000);
+    cy.wait(5000);
     cy.get(".duplicate-notification").should("exist");
   });
 
@@ -144,9 +145,9 @@ describe("Upload a file", function () {
     //create a unique name
     const randomName = Math.random().toString(36).substring(2, 7);
     cy.addFolder(randomName);
-    cy.wait(10000);
+    cy.wait(5000);
     cy.get('[data-testid="upload-file"]').click({ force: true });
-    cy.wait(10000);
+    cy.wait(5000);
 
     //trying to avoid '0x0' effective width and height error and type into input field
     cy.contains("Folder name")
@@ -154,15 +155,18 @@ describe("Upload a file", function () {
       .find("input")
       .type(randomName, { force: true });
 
-    cy.get('[role="listbox"]').find("li").contains(randomName).click();
-    cy.wait(15000);
+    cy.get('[role="listbox"]')
+      .find("li")
+      .contains(randomName)
+      .click({ force: true });
+    cy.wait(8000);
 
     // upload the first fixture file
     cy.get(".upload-btn-wrapper")
       .find("input")
       .invoke("show")
       .selectFile("cypress/fixtures/text-files/text-file.txt");
-    cy.wait(25000);
+    cy.wait(8000);
 
     //upload another fixture file
     cy.get(".upload-btn-wrapper")
@@ -170,52 +174,95 @@ describe("Upload a file", function () {
       .invoke("show")
       .selectFile("cypress/fixtures/text-files/text-file-v2.txt");
 
-    cy.wait(20000);
-    cy.get(".upload-card > c-card-actions.hydrated > :nth-child(2)").click();
-    cy.wait(20000);
+    cy.wait(8000);
+    cy.get(".upload-card > c-card-actions.hydrated > :nth-child(2)").click({
+      force: true,
+    });
+    cy.wait(8000);
 
     //close the modal
-    cy.get(".link-underline").click();
+    cy.get(".link-underline").click({ force: true });
 
     //check the success
     cy.reload();
-    cy.wait(13000);
+    cy.wait(5000);
     cy.contains("text-file").should("exist");
     cy.contains("text-file-v2").should("exist");
 
-    //TODO delete with checkboxes
-
     //delete the first file
-    cy.contains("text-file-v2")
-      .parent()
-      .parent()
-      .find("td")
-      .eq(5)
-      .find("button")
-      .eq(2)
-      .click({ force: true });
-    cy.get("c-alert.hydrated > c-card-actions.hydrated > :nth-child(2)").click({
-      force: true,
-    });
+    cy.deleteFileCheckbox("text-file-v2");
 
-    cy.wait(10000);
+    cy.wait(5000);
 
     //delete the second file
-    cy.contains("text-file")
-      .parent()
-      .parent()
-      .find("td")
-      .eq(5)
-      .find("button")
-      .eq(2)
-      .click({ force: true });
-    cy.get("c-alert.hydrated > c-card-actions.hydrated > :nth-child(2)").click({
+    cy.deleteFileCheckbox("text-file");
+
+    cy.wait(8000);
+    cy.contains("This folder has no content.").should("exist");
+  });
+
+  it("Several files with different names can be uploaded to a folder one by one", () => {
+    //create a unique name
+    const folderName = Math.random().toString(36).substring(2, 7);
+    cy.addFolder(folderName);
+    cy.wait(5000);
+    cy.searchFolder(folderName);
+    cy.get(".media-content").contains(folderName).click({ force: true });
+    cy.wait(3000);
+
+    //press upload button from folder
+    cy.get('[data-testid="upload-file"]').click({ force: true });
+    cy.wait(5000);
+
+    // upload the first fixture file
+    cy.get(".upload-btn-wrapper")
+      .find("input")
+      .invoke("show")
+      .selectFile("cypress/fixtures/text-files/text-file.txt");
+    cy.wait(8000);
+
+    cy.get(".upload-card > c-card-actions.hydrated > :nth-child(2)").click({
       force: true,
     });
+    cy.wait(8000);
 
-    cy.wait(15000);
+    //close the modal
+    cy.get(".toast-main > c-button.hydrated").click({ force: true });
+
+    //press upload button from folder
+    cy.get('[data-testid="upload-file"]').click({ force: true });
+    cy.wait(5000);
+
+    //upload another fixture file
+    cy.get(".upload-btn-wrapper")
+      .find("input")
+      .invoke("show")
+      .selectFile("cypress/fixtures/text-files/text-file-v2.txt");
+
+    cy.wait(8000);
+    cy.get(".upload-card > c-card-actions.hydrated > :nth-child(2)").click({
+      force: true,
+    });
+    cy.wait(8000);
+
+    //close the modal
+    cy.get(".toast-main > c-button.hydrated").click({ force: true });
+
+    //check the success
+    cy.reload();
+    cy.wait(5000);
+    cy.contains("text-file").should("exist");
+    cy.contains("text-file-v2").should("exist");
+
+    //delete the first file
+    cy.deleteFileCheckbox("text-file-v2");
+
+    cy.wait(5000);
+
+    //delete the second file
+    cy.deleteFileCheckbox("text-file");
+
+    cy.wait(8000);
     cy.contains("This folder has no content.").should("exist");
   });
 });
-
-//TODO: test for uploading files with the same path one by one (two different upload modalsc)