+++ tests/cypress/support/commands.js
@@ -130,6 +130,23 @@ Cypress.Commands.add("deleteFile", (fileName) => {
     .eq(2)
     .find("button")
     .click();
+  cy.get("c-alert.hydrated > c-card-actions.hydrated > :nth-child(2)").click({
+    force: true,
+  });
+});
+
+Cypress.Commands.add("deleteFileCheckbox", (fileName) => {
+  cy.contains(fileName)
+    .parent()
+    .parent()
+    .find("td")
+    .eq(5)
+    .find("button")
+    .eq(2)
+    .click({ force: true });
+  cy.get("c-alert.hydrated > c-card-actions.hydrated > :nth-child(2)").click({
+    force: true,
+  });
 });
 
 Cypress.Commands.add("searchFolder", (folderName) => {