+++ CHANGELOG.md
@@ -69,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (GL #999) Add three level sharing with view option
 - (GL #947) Added cypress tests for login, adding and showing containers, uploading files
 - (GL #1047) Added redis and vault monitoring to `/health` endpoint
+- (GL #947) Added cypress tests for sharing folders
 
 ### Changed